// src/components/cards/highlight-card/highlight-card-type.ts
interface HighlightCardType {
  hasVideo: boolean;
  backgroundUrl: string;
  logoUrl: string;
  description: string;
  titleType: string;
  mediaId: boolean;
  titleId: string;
  type: string;
  broadcast: boolean;
  headline: string;
  headlineImage: string;
  serviceId: number;
  idSalePage: string;
  position: number;
}

export default HighlightCardType;